+++ samples/CSharpStyleValidator.Demo/GlobalUsings.cs
@@ -1,5 +1,7 @@
 // Copyright © 2026 DevAM. Licensed under the MIT License. See LICENSE in the repository root.
 
 global using System;
+global using System.Collections.Generic;
+global using System.Text;
 global using System.Threading.Tasks;
 global using Microsoft.CodeAnalysis;